+++ myocamlbuild.ml
@@ -68,30 +68,45 @@ let lib_with_clib ~lib ~clib ~has_lib ~src_dir ~stublib =
     ocaml_lib ~tag_name:use_lib ~dir:src_dir (strf "%s/%s" src_dir lib)
   end
 
-(* tsdl_const.ml generation. *)
+let ctypes_stub_gen () =
+  (* Types stubs generators *)
+  (* Generate stubs. Steps 1, 2, & 3 of Makefile (1 & 2 via built-in rules).
+     ML -> C *)
+  rule "cstubs: x_c_gen.native -> x_stubs_gen.c"
+    ~dep:"%_c_gen.native"
+    ~prod:"%_stubs_gen.c"
+    (fun env _build -> Cmd (A (env "./%_c_gen.native")));
 
-let sdl_consts_build () =
-  let obj s = match !Ocamlbuild_plugin.Options.ext_obj with
-  | "" -> s ^ ".o"
-  | x -> s ^ "." ^ x
-  in
-  dep [ "link"; "ocaml"; "link_consts_stub" ] [ obj "support/consts_stub" ];
-  dep [ "sdl_consts" ] [ "src/tsdl_consts.ml" ];
-  rule "sdl_consts: consts.byte -> tsdl_consts.ml"
-    ~dep:"support/consts.byte"
-    ~prod:"src/tsdl_consts.ml"
-    begin fun env build ->
-      let enums = env "support/consts.byte" in
-      let prod = env "src/tsdl_consts.ml" in
-      Cmd (S [A enums; A prod])
-    end;
-;;
+  (* Step 4. OCamlbuild (nor ocamlc/ocamlopt) has a built in rule for
+     linking executables from C. Call out to 'cc'. *)
+  rule "stub_gen 1: x_stubs_gen.o -> x_stubs_gen"
+    ~dep:"%_stubs_gen.o"
+    ~prod:"%_stubs_gen"
+    (fun env _build ->
+      Cmd (S [ A "cc"; A "-o"; A (env "%_stubs_gen"); A (env "%_stubs_gen.o") ]));
+
+  (* Step 5. Generate ml stubs.  C -> ML  *)
+  rule "stubs_gen 2: x_stubs_gen -> x_stubs.ml"
+    ~dep:"support/%_stubs_gen"
+    ~prod:"src/%_stubs.ml"
+    (fun env _build -> Cmd (S[A (env "support/%_stubs_gen"); Sh">"; A (env "src/%_stubs.ml")]));
+
+  (* Functions stubs generators *)
+  rule "cstubs functions: x_generator.native -> x_stubs.c"
+    ~dep:"support/%_generator.native"
+    ~prod:"src/%_stubs.c"
+    (fun env _build -> Cmd (S[A (env "./support/%_generator.native"); A "c"]));
+
+ rule "mlstubs functions: x_generator.native -> x_generated.ml"
+    ~dep:"support/%_generator.native"
+    ~prod:"src/%_generated.ml"
+    (fun env _build -> Cmd (S[A (env "./support/%_generator.native"); A "ml"]))
 
 let () =
   dispatch begin function
   | After_rules ->
       lib_with_clib ~lib:"tsdl" ~clib:"sdl2" ~has_lib:"-DHAS_SDL2"
         ~src_dir:"src" ~stublib:"tsdl_stubs";
-      sdl_consts_build ()
+      ctypes_stub_gen ()
   | _ -> ()
   end