+++ CHANGES.md
@@ -1,5 +1,7 @@
 - Fix bug in `Message_box` color scheme handling.
   Thanks to Matthieu Dubuget for the patch (#105).
+- Remove legacy function `get_revision_number` (`get_revision` should
+  be used instead).
 
 v1.1.0 2024-09-12 Zagreb
 ------------------------